requirements.txt
uvicorn==0.18.3
fastapi==0.85.0
SQLAlchemy==1.4.41
Jinja2==3.1.2
python-dotenv==0.21.0
coloredlogs==15.0.1
asyncpg==0.26.0
ormar==0.11.3
psycopg2-binary==2.9.3